+++ frontend/src/service/playlist-service.ts
@@ -12,12 +12,16 @@ export default class PlaylistService {
         return new Observable((obs) =>
             this.playlistApiService.getPlaylist(url).pipe(first()).subscribe({
                 next: (pl) => {
-                    let cnt = 0;
-                    pl.forEach(g => {
-                        g.id = ++cnt;
-                        g.channels.forEach(c => c.id = ++cnt);
-                    })
-                    obs.next(pl);
+                    if (pl) {
+                        let cnt = 0;
+                        pl.forEach(g => {
+                            g.id = ++cnt;
+                            g.channels.forEach(c => c.id = ++cnt);
+                        })
+                        obs.next(pl);
+                    } else {
+                        obs.error("Could not download playlist");
+                    }
                 },
                 error: (e) => obs.error(e),
                 complete: () => obs.complete(),